Move depsHash into project.pbxproj
We were previously writing `depsHash` into `XCHammerAssets/depsHash` in
an atomic transaction.

Despite this, it's possible in rare circumstances that this file gets
out of sync with the Xcode project. For example, if Xcode decides to
write the in memory representation of the project ontop of our changes.

This patch associates the `project.pbxproj` with a given `depsHash` by
writing it into a user defined build setting. We attempt to read it from
the env if it exists, otherwise, quickly parse it from the project.

Additionally, I took the liberty to address #1 due for stability
reasons. If the file system changes during generation ( unfortunately
possible ), we will have saved the hash that we built the project with.

Fixes #2
Fixes #1