Bazel integration Pass 1
This patch adds the ability to build bazel targets directly in the
XCHammer projects.

Build targets are generated only for the labels that are explicitly
added to the config.

Add an integration test to make sure we can build via custom schemes.

Additionally, I added runscript templating. We need to dynamically
configure Bazel invocations based on the env. This is useful for things
like static analysis or other custom behaviors.

Debugging uses lldbinit for reproducibility and speed. In pratice, I map
to a stablized directory which enables debugging of remotely built and
cached of object files without dsyms.

The feature is optional since it relies on custom instrumented behavior
for debugging - the debugger will not work without the internal
toolchain. Eventually, we should consider integrating a sensible
default like include the toolchain with XCHammer.