Generalize Genrule Query
Previously, we hardcoded known rule types to genrules. This is quite
limiting and requires a user to change XCHammer.

Now, the query include rules of tag 'xchammer'. The user can include
arbitrary types of rules to be built for Xcode usage.  These files are
assumed to be requsites of Xcode.

This is a prelude to #46 which will include a followup to run `build`
before generation and corresponding logic to map paths.