Move Samples to Goldmaster Part 1
There is a lot of noise in the sample project, which was never meant to
be checked in.

This commit adds stable, goldmaster files to integration tests to reduce
some of the noise in code reviews.

If it isn't reasonable to make the output 100% stable, then we should
diff projects with XcodeGen

Resolves #13