Generate XCHammer .xcodeproj with XCHammer
This commit adds the ability to build XCHammer's Xcode project with
XCHammer.

Additionally change `make workspace` to use this instead of SPM.
Incrmental builds aren't 100% out of the box, but we should add that to
Bazel and additionally support building the XCHammer Xcode project with
XCHammer.

The legacy generation command is left until incremental builds are
working.

Known issues with Swift and XCHammer:
- includes in external
- module_name propagation
- `BUILD` file naming edge cases in resources