Adopt subset of XcodeGen.TargetScheme (#130)
This PR adopts a subset of XcodeGen's Target Scheme. Mainly, it allows
people to select scripts to run during schemes.

The TargetSchemes are identical and shared across Bazel / Xcode targets

For now, it seems OK to have this as there are no idioms in Bazel to
achieve the same result.