Import XCConfigExporter sources
In order to sync up Xcode diagnostics with Bazel, we extract desired
compiler flags from Xcode via an xcconfig file by reading the compiler
commands.

This code was instrumented using internal data strucutres of Xcode 9
which have since changed.

For the most part, this was a 1 off program but the approach here seemed
like the only reliable way to do this at the time.

Future Work:
- migrate compilation DB to log parsing approach:
https://github.com/jerrymarino/XcodeCompilationDatabase
- set this up with Bazel and XCHammer
- fully integrate Bazel ( XCFocus ) to Xcode and delete this.

This approch is useful for making Xcode builds work like Bazels