Experimental xcode project rule
Implement a rule which is a consumer of the aspect. This rule simply
runs XCHammer project generation logic, but letting Bazel drive.

It simplifies and improves a number of things:
- Remove the need for DepsHash. This simply uses Bazel to determine that
- Remove parallel generator. Uses Bazel to paralleize
- Remote caching ( in some regards ) I'd like to build the projects on
master and pull from the cache locally.
- Keeping generation incremental / analysis cache hot. Currently bazel
discards analysis cache which can be slow. I'd like to run this during
every build with the same arguments as the builds
e.g. bazel build ios-app ios-app.xcodeproj
- Error reporting and diags. This removes layers of abstractions to get error
messages in deeply nested deps / subprocesses

Tradeoffs:
- Output writing is not streamed. This may be OK if it's fast ( < 10s )
- Ability to govern Bazel as running recursivly is complex?
- Need to use this as an input to some non-hermetic rule to run after
the generator. Projects currently can't live in `bazel-bin`

TODOs are added inline for what needs to change