Xcode test Bazel built test bundles
This PR adds the ability to run Bazel built test bundles with Xcode.

There's a few basic additions, like adding test hosts.

When there's a test host, xcode first builds the test host, and then
bazel builds the test host. Finally, it installs the test bundle into
the app.

Support for "Focused" testing

In order for the focused test buttons to pop up, we need to add source
files to Xcode. We don't actually want to build the source files, so the
CC is stubbed out with a compiler that produces a dummy object file.

These "dots" are driven by the indexer, and the files must have the
correct settings as well.