[Focus] Support transitive source file inclusion (#209)
* [Focus] Support transitive source file inclusion

- use `generateXcodeSchemes` as a switch to determine Xcode dependency
requirements. We need to test this more for swift.
- use `generateTransitiveXcodeTargets` to switch target inclusion. This
allows the users to easily select targets

For many Xcode projects, we have a lot of targets and want to default to
pull most of them in. A user sets `generateTransitiveXcodeTargets=true`,
and then includes/excludes sources based on paths

Additionally update goldmaster and fix an issue in the `Makefile`