Integrate XCBuildKit Xcode Progress bar (#207)
This PR sets up the Xcode progress bar in XCHammer using XCBuildKit.

Simply put, we use the BazelBuildService for now, which allows us to show
progress to the user. In the future, XCHammer may implement it's own build
service to run Xcode transparently.

This command runs the installer from XCBuildKit as an ad-hoc subcommand. It has
a very fast noop time and is intended to be ran as part of project generation.
Eventually we'll add it into gen/builds and not have an ad-hoc invocation.