Improve handling of Swift under Bazel builds
First, it removes transitive module map collection. This is no longer
needed and produces a different result.

Next, It improves the handling of copts in general to adopt the calling
`expand_location` inside of the aspect. In PodToBUILD, we just landed a
series of changes to build the top 50 swift cocoapods. Like other usages
of building mixed module swift libs, it uses `$(execpath` to pass an
internal module map via fmodule-map-file=. The reference Tulsi PR
resides here: https://github.com/bazelbuild/tulsi/pull/157 and I bumped
XCHammer to include this here
https://github.com/pinterest/xchammer/pull/243

Finally, it adds installation of swiftmodules and corresponding include
paths. This makes it easy for sourcekit to pick them up. Additionally, it
creates a swift stub to prevent Xcode from building swift files. This is
conditional for pure Xcode builds, as we build the swift modules with
Xcode in that case.

Additionally, I fixed an issue with `BazelExtensions` caught running
`make workspace` which is now part of the CI