Bump Tulsi to HEAD (#247)
We've hit a few other issues when upgrading to Bazel 3.4.1 that seemed
intermittent before.

Under XCBuild it will conditionally install the XCTest dylibs and
other dylibs only if the bundle has changed. Under XCBuild the
result is that the dylibs are stripped out of the bundle from the
previous build. This change also sync bundles directly to the output
location if they exist.

Please find more info in the linked commit