Stop gap for "Could not attach to pid:" (#248)
* Stop gap for "Could not attach to pid:"

On my encatation of macOS Catalina, XCBuild, and Xcode 11, the simulator
intermittently fails with the error "Could not attach to pid:"

It turns out that there's 2 symptoms

1: the binary doesn't launch
2: lldb fails to attach w/o get-task-allow

For now, the workaround allows XCBuild to code sign if needed. Longer
term, we shouldn't really need this but folks often have to add the
setting manually.

Additionally, streamlines some of the code for coding platform specific keys.