Upload binary release artfact on all builds
This PR adds the ability to upload release artifact via XCHammer's github
CI.

On each build, we upload the binary release artifact `xchammer.zip`
which can be pulled into a bazel `WORKSPACE` with `http_archive` and
then used as expected.

At first, I'd considered creating a release on each commit but that has
a few complexities in github actions e.g. can't create a release w/o a
tag, and would end up annoying watchers of the repository with a release
email on each commit landed. Perhaps there could be a better way to navigate
to these artifact URLs in the github UI longer term.