Add `compilation_mode` into XCHammer's own xcodeproj
By default this is set to `fastbuild` and debug info isn't set or
working. This PR uses existing config systems to set the
`compilation_mode`

Additionally document current config possibilities which don't really
handle this by default.

After running into this in XCHammer it's pretty clear to me that we
should have a sensible default to set the `compilation_mode` by default.
Perhaps a followup could consider adding defaults. I realized we don't
have this issue in the test sample because it sets the default
`compilation_mode` to `dbg` in the `.bazelrc`